Moved all value classes out of StepConverters.
