+++ src/main/scala/scala/compat/java8/StepConverters.scala
@@ -8,381 +8,6 @@ import scala.compat.java8.runtime._
 package converterImpls {
   import Stepper._
 
-  final class RichArrayAnyCanStep[A](private val underlying: Array[A]) extends AnyVal with MakesAnyStepper[A] {
-    @inline def stepper: AnyStepper[A] with EfficientSubstep = new StepsAnyArray[A](underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayObjectCanStep[A <: Object](private val underlying: Array[A]) extends AnyVal with MakesAnyStepper[A] {
-    @inline def stepper: AnyStepper[A] with EfficientSubstep = new StepsObjectArray[A](underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayUnitCanStep(private val underlying: Array[Unit]) extends AnyVal with MakesAnyStepper[Unit] {
-    @inline def stepper: AnyStepper[Unit] with EfficientSubstep = new StepsUnitArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayBooleanCanStep(private val underlying: Array[Boolean]) extends AnyVal with MakesAnyStepper[Boolean] {
-    @inline def stepper: AnyStepper[Boolean] with EfficientSubstep = new StepsBoxedBooleanArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayByteCanStep(private val underlying: Array[Byte]) extends AnyVal with MakesAnyStepper[Byte] {
-    @inline def stepper: AnyStepper[Byte] with EfficientSubstep = new StepsBoxedByteArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayCharCanStep(private val underlying: Array[Char]) extends AnyVal with MakesAnyStepper[Char] {
-    @inline def stepper: AnyStepper[Char] with EfficientSubstep = new StepsBoxedCharArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayShortCanStep(private val underlying: Array[Short]) extends AnyVal with MakesAnyStepper[Short] {
-    @inline def stepper: AnyStepper[Short] with EfficientSubstep = new StepsBoxedShortArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayFloatCanStep(private val underlying: Array[Float]) extends AnyVal with MakesAnyStepper[Float] {
-    @inline def stepper: AnyStepper[Float] with EfficientSubstep = new StepsBoxedFloatArray(underlying, 0, underlying.length)
-  }
-
-  final class RichIndexedSeqCanStep[A](private val underlying: collection.IndexedSeqLike[A, _]) extends AnyVal with MakesAnyStepper[A] {
-    @inline def stepper: AnyStepper[A] with EfficientSubstep = new StepsAnyIndexedSeq[A](underlying, 0, underlying.length)
-  }
-  
-  final class RichDoubleIndexedSeqCanStep[CC <: collection.IndexedSeqLike[Double, _]](private val underlying: CC) extends AnyVal with MakesDoubleStepper {
-    @inline def stepper: DoubleStepper with EfficientSubstep = new StepsDoubleIndexedSeq[CC](underlying, 0, underlying.length)
-  }
-  
-  final class RichIntIndexedSeqCanStep[CC <: collection.IndexedSeqLike[Int, _]](private val underlying: CC) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepsIntIndexedSeq[CC](underlying, 0, underlying.length)
-  }
-  
-  final class RichLongIndexedSeqCanStep[CC <: collection.IndexedSeqLike[Long, _]](private val underlying: CC) extends AnyVal with MakesLongStepper {
-    @inline def stepper: LongStepper with EfficientSubstep = new StepsLongIndexedSeq[CC](underlying, 0, underlying.length)
-  }
-
-  final class RichLinearSeqCanStep[A, CC >: Null <: collection.LinearSeqLike[A, CC]](private val underlying: CC) extends AnyVal with MakesAnySeqStepper[A] {
-    @inline def stepper: AnyStepper[A] = new StepsAnyLinearSeq[A, CC](underlying, Long.MaxValue)
-  }
-
-  final class RichDoubleLinearSeqCanStep[CC >: Null <: collection.LinearSeqLike[Double, CC]](private val underlying: CC) extends AnyVal with MakesDoubleSeqStepper {
-    @inline def stepper: DoubleStepper = new StepsDoubleLinearSeq[CC](underlying, Long.MaxValue)
-  }
-
-  final class RichIntLinearSeqCanStep[CC >: Null <: collection.LinearSeqLike[Int, CC]](private val underlying: CC) extends AnyVal with MakesIntSeqStepper {
-    @inline def stepper: IntStepper = new StepsIntLinearSeq[CC](underlying, Long.MaxValue)
-  }
-
-  final class RichLongLinearSeqCanStep[CC >: Null <: collection.LinearSeqLike[Long, CC]](private val underlying: CC) extends AnyVal with MakesLongSeqStepper {
-    @inline def stepper: LongStepper = new StepsLongLinearSeq[CC](underlying, Long.MaxValue)
-  }
-
-  final class RichIteratorCanStep[A](private val underlying: Iterator[A]) extends AnyVal with MakesAnySeqStepper[A] {
-    @inline def stepper: AnyStepper[A] = new StepsAnyIterator[A](underlying)
-  }
-
-  final class RichDoubleIteratorCanStep(private val underlying: Iterator[Double]) extends AnyVal with MakesDoubleSeqStepper {
-    @inline def stepper: DoubleStepper = new StepsDoubleIterator(underlying)
-  }
-
-  final class RichIntIteratorCanStep(private val underlying: Iterator[Int]) extends AnyVal with MakesIntSeqStepper {
-    @inline def stepper: IntStepper = new StepsIntIterator(underlying)
-  }
-
-  final class RichLongIteratorCanStep(private val underlying: Iterator[Long]) extends AnyVal with MakesLongSeqStepper {
-    @inline def stepper: LongStepper = new StepsLongIterator(underlying)
-  }
-
-  final class RichNumericRangeCanStep[T](private val underlying: collection.immutable.NumericRange[T]) extends AnyVal with MakesAnyStepper[T] {
-    @inline def stepper: AnyStepper[T] with EfficientSubstep = new StepsAnyNumericRange[T](underlying, 0, underlying.length)
-  }
-
-  final class RichVectorCanStep[A](private val underlying: Vector[A]) extends AnyVal with MakesAnyStepper[A] {
-    @inline def stepper: AnyStepper[A] with EfficientSubstep = new StepsAnyVector[A](underlying, 0, underlying.length)
-  }
-
-  final class RichFlatHashTableCanStep[A](private val underlying: collection.mutable.FlatHashTable[A]) extends AnyVal with MakesAnyStepper[A] {
-    @inline def stepper: AnyStepper[A] with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable(underlying)
-      new StepsAnyFlatHashTable(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichDoubleFlatHashTableCanStep(private val underlying: collection.mutable.FlatHashTable[Double]) extends AnyVal with MakesDoubleStepper {
-    @inline def stepper: DoubleStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable(underlying)
-      new StepsDoubleFlatHashTable(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichIntFlatHashTableCanStep(private val underlying: collection.mutable.FlatHashTable[Int]) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable(underlying)
-      new StepsIntFlatHashTable(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichLongFlatHashTableCanStep(private val underlying: collection.mutable.FlatHashTable[Long]) extends AnyVal with MakesLongStepper {
-    @inline def stepper: LongStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable(underlying)
-      new StepsLongFlatHashTable(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichHashTableKeyCanStep[K, HE >: Null <: collection.mutable.HashEntry[K, HE]](private val underlying: collection.mutable.HashTable[K, HE])
-  extends AnyVal with MakesAnyKeyStepper[K] {
-    @inline def keyStepper: AnyStepper[K] with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, HE](underlying)
-      new StepsAnyHashTableKey(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichHashTableDoubleKeyCanStep[HE >: Null <: collection.mutable.HashEntry[Double, HE]](private val underlying: collection.mutable.HashTable[Double, HE])
-  extends AnyVal with MakesDoubleKeyStepper {
-    @inline def keyStepper: DoubleStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[Double, HE](underlying)
-      new StepsDoubleHashTableKey(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichHashTableIntKeyCanStep[HE >: Null <: collection.mutable.HashEntry[Int, HE]](private val underlying: collection.mutable.HashTable[Int, HE])
-  extends AnyVal with MakesIntKeyStepper {
-    @inline def keyStepper: IntStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[Int, HE](underlying)
-      new StepsIntHashTableKey(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichHashTableLongKeyCanStep[HE >: Null <: collection.mutable.HashEntry[Long, HE]](private val underlying: collection.mutable.HashTable[Long, HE])
-  extends AnyVal with MakesLongKeyStepper {
-    @inline def keyStepper: LongStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[Long, HE](underlying)
-      new StepsLongHashTableKey(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichDefaultHashTableCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, V]])
-  extends AnyVal with MakesAnyStepper[(K, V)] {
-    @inline def stepper: AnyStepper[(K,V)] with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, V]](underlying)
-      new StepsAnyDefaultHashTable(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichDefaultHashTableValueCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, V]])
-  extends AnyVal with MakesAnyValueStepper[V] {
-    @inline def valueStepper: AnyStepper[V] with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, V]](underlying)
-      new StepsAnyDefaultHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichDefaultHashTableDoubleValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, Double]])
-  extends AnyVal with MakesDoubleValueStepper {
-    @inline def valueStepper: DoubleStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, Double]](underlying)
-      new StepsDoubleDefaultHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichDefaultHashTableIntValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, Int]])
-  extends AnyVal with MakesIntValueStepper {
-    @inline def valueStepper: IntStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, Int]](underlying)
-      new StepsIntDefaultHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichDefaultHashTableLongValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, Long]])
-  extends AnyVal with MakesLongValueStepper {
-    @inline def valueStepper: LongStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, Long]](underlying)
-      new StepsLongDefaultHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-    
-  final class RichLinkedHashTableCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, V]])
-  extends AnyVal with MakesAnyStepper[(K,V)] {
-    @inline def stepper: AnyStepper[(K,V)] with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, V]](underlying)
-      new StepsAnyLinkedHashTable(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichLinkedHashTableValueCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, V]])
-  extends AnyVal with MakesAnyValueStepper[V] {
-    @inline def valueStepper: AnyStepper[V] with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, V]](underlying)
-      new StepsAnyLinkedHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichLinkedHashTableDoubleValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, Double]])
-  extends AnyVal with MakesDoubleValueStepper {
-    @inline def valueStepper: DoubleStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, Double]](underlying)
-      new StepsDoubleLinkedHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichLinkedHashTableIntValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, Int]])
-  extends AnyVal with MakesIntValueStepper {
-    @inline def valueStepper: IntStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, Int]](underlying)
-      new StepsIntLinkedHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-  
-  final class RichLinkedHashTableLongValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, Long]])
-  extends AnyVal with MakesLongValueStepper {
-    @inline def valueStepper: LongStepper with EfficientSubstep = {
-      val tbl = runtime.CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, Long]](underlying)
-      new StepsLongLinkedHashTableValue(tbl, 0, tbl.length)
-    }
-  }
-
-  final class RichImmHashMapCanStep[K, V](private val underlying: collection.immutable.HashMap[K, V])
-  extends AnyVal with MakesAnyStepper[(K, V)] with MakesAnyKeyStepper[K] with MakesAnyValueStepper[V] {
-    @inline def stepper: AnyStepper[(K, V)] with EfficientSubstep = new StepsAnyImmHashMap[K, V](underlying, 0, underlying.size)
-    @inline def keyStepper: AnyStepper[K] with EfficientSubstep = new StepsAnyImmHashMapKey[K, V](underlying, 0, underlying.size)
-    @inline def valueStepper: AnyStepper[V] with EfficientSubstep = new StepsAnyImmHashMapValue[K, V](underlying, 0, underlying.size)
-  }
-
-  final class RichImmHashMapDoubleKeyCanStep[V](private val underlying: collection.immutable.HashMap[Double, V]) extends AnyVal with MakesDoubleKeyStepper {
-    @inline def keyStepper: DoubleStepper with EfficientSubstep = new StepsDoubleImmHashMapKey[V](underlying, 0, underlying.size)
-  }
-
-  final class RichImmHashMapDoubleValueCanStep[K](private val underlying: collection.immutable.HashMap[K, Double]) extends AnyVal with MakesDoubleValueStepper {
-    @inline def valueStepper: DoubleStepper with EfficientSubstep = new StepsDoubleImmHashMapValue[K](underlying, 0, underlying.size)
-  }
-    
-  final class RichImmHashMapIntKeyCanStep[V](private val underlying: collection.immutable.HashMap[Int, V]) extends AnyVal with MakesIntKeyStepper {
-    @inline def keyStepper: IntStepper with EfficientSubstep = new StepsIntImmHashMapKey[V](underlying, 0, underlying.size)
-  }
-
-  final class RichImmHashMapIntValueCanStep[K](private val underlying: collection.immutable.HashMap[K, Int]) extends AnyVal with MakesIntValueStepper {
-    @inline def valueStepper: IntStepper with EfficientSubstep = new StepsIntImmHashMapValue[K](underlying, 0, underlying.size)
-  }
-    
-  final class RichImmHashMapLongKeyCanStep[V](private val underlying: collection.immutable.HashMap[Long, V]) extends AnyVal with MakesLongKeyStepper {
-    @inline def keyStepper: LongStepper with EfficientSubstep = new StepsLongImmHashMapKey[V](underlying, 0, underlying.size)
-  }
-
-  final class RichImmHashMapLongValueCanStep[K](private val underlying: collection.immutable.HashMap[K, Long]) extends AnyVal with MakesLongValueStepper {
-    @inline def valueStepper: LongStepper with EfficientSubstep = new StepsLongImmHashMapValue[K](underlying, 0, underlying.size)
-  }
-    
-  final class RichImmHashSetCanStep[A](private val underlying: collection.immutable.HashSet[A]) extends AnyVal with MakesAnyStepper[A] {
-    @inline def stepper: AnyStepper[A] with EfficientSubstep = new StepsAnyImmHashSet(underlying.iterator, underlying.size)
-  }
-
-  final class RichBitSetCanStep(private val underlying: collection.BitSet) extends AnyVal with MakesIntStepper {
-    def stepper: IntStepper with EfficientSubstep = {
-      val bits: Array[Long] = underlying match {
-        case m: collection.mutable.BitSet => runtime.CollectionInternals.getBitSetInternals(m)
-        case n: collection.immutable.BitSet.BitSetN => RichBitSetCanStep.reflectInternalsN(n)
-        case x => x.toBitMask
-      }
-      new StepsIntBitSet(bits, 0, math.min(bits.length*64L, Int.MaxValue).toInt)
-    }
-  }
-  private[java8] object RichBitSetCanStep {
-    private val reflector = classOf[collection.immutable.BitSet.BitSetN].getMethod("elems")
-    def reflectInternalsN(bsn: collection.immutable.BitSet.BitSetN): Array[Long] = reflector.invoke(bsn).asInstanceOf[Array[Long]]
-  }
-
-  final class RichMapCanStep[K, V](private val underlying: collection.Map[K, V]) extends AnyVal with MakesAnyKeySeqStepper[K] with MakesAnyValueSeqStepper[V] {
-    def keyStepper: AnyStepper[K] = new StepsAnyIterator[K](underlying.keysIterator)
-    def valueStepper: AnyStepper[V] = new StepsAnyIterator[V](underlying.valuesIterator)
-  }
-
-  final class RichDoubleKeyMapCanStep[V](private val underlying: collection.Map[Double, V]) extends AnyVal with MakesDoubleKeySeqStepper {
-    def keyStepper: DoubleStepper = new StepsDoubleIterator(underlying.keysIterator)
-  }
-
-  final class RichDoubleValueMapCanStep[K](private val underlying: collection.Map[K, Double]) extends AnyVal with MakesDoubleValueSeqStepper {
-    def valueStepper: DoubleStepper = new StepsDoubleIterator(underlying.valuesIterator)
-  }
-
-  final class RichIntKeyMapCanStep[V](private val underlying: collection.Map[Int, V]) extends AnyVal with MakesIntKeySeqStepper {
-    def keyStepper: IntStepper = new StepsIntIterator(underlying.keysIterator)
-  }
-
-  final class RichIntValueMapCanStep[K](private val underlying: collection.Map[K, Int]) extends AnyVal with MakesIntValueSeqStepper {
-    def valueStepper: IntStepper = new StepsIntIterator(underlying.valuesIterator)
-  }
-
-  final class RichLongKeyMapCanStep[V](private val underlying: collection.Map[Long, V]) extends AnyVal with MakesLongKeySeqStepper {
-    def keyStepper: LongStepper = new StepsLongIterator(underlying.keysIterator)
-  }
-
-  final class RichLongValueMapCanStep[K](private val underlying: collection.Map[K, Long]) extends AnyVal with MakesLongValueSeqStepper {
-    def valueStepper: LongStepper = new StepsLongIterator(underlying.valuesIterator)
-  }
-
-  final class RichIterableCanStep[A](private val underlying: Iterable[A]) extends AnyVal with MakesAnySeqStepper[A] {
-    @inline def stepper: AnyStepper[A] = new StepsAnyIterator[A](underlying.iterator)
-  }
-
-  final class RichDoubleIterableCanStep(private val underlying: Iterable[Double]) extends AnyVal with MakesDoubleSeqStepper {
-    @inline def stepper: DoubleStepper = new StepsDoubleIterator(underlying.iterator)
-  }
-
-  final class RichIntIterableCanStep(private val underlying: Iterable[Int]) extends AnyVal with MakesIntSeqStepper {
-    @inline def stepper: IntStepper = new StepsIntIterator(underlying.iterator)
-  }
-
-  final class RichLongIterableCanStep(private val underlying: Iterable[Long]) extends AnyVal with MakesLongSeqStepper {
-    @inline def stepper: LongStepper = new StepsLongIterator(underlying.iterator)
-  }
-
-  final class RichArrayDoubleCanStep(private val underlying: Array[Double]) extends AnyVal with MakesDoubleStepper {
-    @inline def stepper: DoubleStepper with EfficientSubstep with EfficientSubstep = new StepsDoubleArray(underlying, 0, underlying.length)
-  }
-
-  final class RichArrayIntCanStep(private val underlying: Array[Int]) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepsIntArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichArrayLongCanStep(private val underlying: Array[Long]) extends AnyVal with MakesLongStepper {
-    @inline def stepper: LongStepper with EfficientSubstep = new StepsLongArray(underlying, 0, underlying.length)
-  }
-  
-  final class RichIntNumericRangeCanStep(private val underlying: collection.immutable.NumericRange[Int]) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepsIntNumericRange(underlying, 0, underlying.length)
-  }
-
-  final class RichLongNumericRangeCanStep(private val underlying: collection.immutable.NumericRange[Long]) extends AnyVal with MakesLongStepper {
-    @inline def stepper: LongStepper with EfficientSubstep = new StepsLongNumericRange(underlying, 0, underlying.length)
-  }
-
-  final class RichRangeCanStep(private val underlying: Range) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepsIntRange(underlying, 0, underlying.length)
-  }  
-
-  final class RichDoubleVectorCanStep[A](private val underlying: Vector[Double]) extends AnyVal with MakesDoubleStepper {
-    @inline def stepper: DoubleStepper with EfficientSubstep = new StepsDoubleVector(underlying, 0, underlying.length)
-  }
-
-  final class RichIntVectorCanStep[A](private val underlying: Vector[Int]) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepsIntVector(underlying, 0, underlying.length)
-  }
-
-  final class RichLongVectorCanStep[A](private val underlying: Vector[Long]) extends AnyVal with MakesLongStepper {
-    @inline def stepper: LongStepper with EfficientSubstep = new StepsLongVector(underlying, 0, underlying.length)
-  }
-
-  final class RichDoubleHashSetCanStep(private val underlying: collection.immutable.HashSet[Double]) extends AnyVal with MakesDoubleStepper {
-    @inline def stepper: DoubleStepper with EfficientSubstep = new StepsDoubleImmHashSet(underlying.iterator, underlying.size)
-  }
-
-  final class RichIntHashSetCanStep(private val underlying: collection.immutable.HashSet[Int]) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepsIntImmHashSet(underlying.iterator, underlying.size)
-  }
-
-  final class RichLongHashSetCanStep(private val underlying: collection.immutable.HashSet[Long]) extends AnyVal with MakesLongStepper {
-    @inline def stepper: LongStepper with EfficientSubstep = new StepsLongImmHashSet(underlying.iterator, underlying.size)
-  }
-
-  final class RichStringCanStep(private val underlying: String) extends AnyVal with MakesIntStepper {
-    @inline def stepper: IntStepper with EfficientSubstep = new StepperStringCodePoint(underlying, 0, underlying.length)
-  }    
-
   trait Priority7StepConverters {
     implicit def richIterableCanStep[A](underlying: Iterable[A]) = new RichIterableCanStep(underlying)
   }
@@ -505,68 +130,6 @@ package converterImpls {
     implicit def richStringCanStep(underlying: String) = new RichStringCanStep(underlying)
   }
 
-  final class CollectionCanAccumulate[A](private val underlying: TraversableOnce[A]) extends AnyVal {
-    def accumulate: Accumulator[A] = {
-      val a = new Accumulator[A]
-      underlying.foreach(a += _)
-      a
-    }
-  }
-
-  final class AccumulateDoubleCollection(private val underlying: TraversableOnce[Double]) extends AnyVal {
-    def accumulate: DoubleAccumulator = {
-      val da = new DoubleAccumulator
-      underlying.foreach(da += _)
-      da
-    }
-  }
-  final class AccumulateIntCollection(private val underlying: TraversableOnce[Int]) extends AnyVal {
-    def accumulate: IntAccumulator = {
-      val da = new IntAccumulator
-      underlying.foreach(da += _)
-      da
-    }
-  }
-  final class AccumulateLongCollection(private val underlying: TraversableOnce[Long]) extends AnyVal {
-    def accumulate: LongAccumulator = {
-      val da = new LongAccumulator
-      underlying.foreach(da += _)
-      da
-    }
-  }
-  final class AccumulateAnyArray[A](private val underlying: Array[A]) extends AnyVal {
-    def accumulate: Accumulator[A] = {
-      val a = new Accumulator[A]
-      var i = 0
-      while (i < underlying.length) { a += underlying(i); i += 1 }
-      a
-    }
-  }
-
-  final class AccumulateDoubleArray(private val underlying: Array[Double]) extends AnyVal {
-    def accumulate: DoubleAccumulator = {
-      val da = new DoubleAccumulator
-      var i = 0
-      while (i < underlying.length) { da += underlying(i); i += 1 }
-      da
-    }
-  }
-  final class AccumulateIntArray(private val underlying: Array[Int]) extends AnyVal {
-    def accumulate: IntAccumulator = {
-      val da = new IntAccumulator
-      var i = 0
-      while (i < underlying.length) { da += underlying(i); i += 1 }
-      da
-    }
-  }
-  final class AccumulateLongArray(private val underlying: Array[Long]) extends AnyVal {
-    def accumulate: LongAccumulator = {
-      val da = new LongAccumulator
-      var i = 0
-      while (i < underlying.length) { da += underlying(i); i += 1 }
-      da
-    }
-  }
 
   trait Priority3AccumulatorConverters {
     implicit def collectionCanAccumulate[A](underlying: TraversableOnce[A]) = new CollectionCanAccumulate[A](underlying)
