+++ src/main/scala/scala/compat/java8/collectionImpl/StepsFlatHashTable.scala
@@ -7,6 +7,10 @@ import scala.compat.java8.runtime._
 
 import Stepper._
 
+/////////////////////////////
+// Stepper implementations //
+/////////////////////////////
+
 private[java8] class StepsAnyFlatHashTable[A](_underlying: Array[AnyRef], _i0: Int, _iN: Int)
 extends StepsLikeGapped[A, StepsAnyFlatHashTable[A]](_underlying, _i0, _iN) {
   def next() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[A]; currentEntry = null; ans }
@@ -31,3 +35,34 @@ extends StepsLongLikeGapped[StepsLongFlatHashTable](_underlying, _i0, _iN) {
   def semiclone(half: Int) = new StepsLongFlatHashTable(underlying, i0, half)    
 }
 
+//////////////////////////
+// Value class adapters //
+//////////////////////////
+
+final class RichFlatHashTableCanStep[A](private val underlying: collection.mutable.FlatHashTable[A]) extends AnyVal with MakesAnyStepper[A] {
+  @inline def stepper: AnyStepper[A] with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable(underlying)
+    new StepsAnyFlatHashTable(tbl, 0, tbl.length)
+  }
+}
+
+final class RichDoubleFlatHashTableCanStep(private val underlying: collection.mutable.FlatHashTable[Double]) extends AnyVal with MakesDoubleStepper {
+  @inline def stepper: DoubleStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable(underlying)
+    new StepsDoubleFlatHashTable(tbl, 0, tbl.length)
+  }
+}
+
+final class RichIntFlatHashTableCanStep(private val underlying: collection.mutable.FlatHashTable[Int]) extends AnyVal with MakesIntStepper {
+  @inline def stepper: IntStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable(underlying)
+    new StepsIntFlatHashTable(tbl, 0, tbl.length)
+  }
+}
+
+final class RichLongFlatHashTableCanStep(private val underlying: collection.mutable.FlatHashTable[Long]) extends AnyVal with MakesLongStepper {
+  @inline def stepper: LongStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable(underlying)
+    new StepsLongFlatHashTable(tbl, 0, tbl.length)
+  }
+}
