+++ src/main/scala/scala/compat/java8/collectionImpl/StepsHashTable.scala
@@ -7,9 +7,11 @@ import scala.compat.java8.runtime._
 
 import Stepper._
 
-////////////////////////////
+/////////////////////////////
+// Stepper implementations //
+/////////////////////////////
+
 // Steppers for keys (type of HashEntry doesn't matter)
-////////////////////////////
 
 private[java8] class StepsAnyHashTableKey[K, HE <: collection.mutable.HashEntry[K, HE]](_underlying: Array[collection.mutable.HashEntry[K, HE]], _i0: Int, _iN: Int)
 extends StepsLikeGapped[K, StepsAnyHashTableKey[K, HE]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
@@ -35,10 +37,8 @@ extends StepsLongLikeGapped[StepsLongHashTableKey[HE]](_underlying.asInstanceOf[
   def semiclone(half: Int) = new StepsLongHashTableKey[HE](underlying.asInstanceOf[Array[collection.mutable.HashEntry[Long, HE]]], i0, half)
 }
 
-////////////////////////////
 // Steppers for entries stored in DefaultEntry HashEntry
 // (both for key-value pair and for values alone)
-////////////////////////////
 
 private[java8] class StepsAnyDefaultHashTable[K, V](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, V]]], _i0: Int, _iN: Int)
 extends StepsLikeGapped[(K, V), StepsAnyDefaultHashTable[K, V]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
@@ -85,10 +85,8 @@ extends StepsLongLikeGapped[StepsLongDefaultHashTableValue[K]](_underlying.asIns
     new StepsLongDefaultHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Long]]]], i0, half)
 }
 
-////////////////////////////
 // Steppers for entries stored in LinkedEntry HashEntry
 // (both for key-value pair and for values alone)
-////////////////////////////
 
 private[java8] class StepsAnyLinkedHashTable[K, V](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, V]]], _i0: Int, _iN: Int)
 extends StepsLikeGapped[(K, V), StepsAnyLinkedHashTable[K, V]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
@@ -135,3 +133,127 @@ extends StepsLongLikeGapped[StepsLongLinkedHashTableValue[K]](_underlying.asInst
     new StepsLongLinkedHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Long]]]], i0, half)
 }
 
+
+//////////////////////////
+// Value class adapters //
+//////////////////////////
+
+// Steppers for keys (type of HashEntry doesn't matter)
+
+final class RichHashTableKeyCanStep[K, HE >: Null <: collection.mutable.HashEntry[K, HE]](private val underlying: collection.mutable.HashTable[K, HE])
+extends AnyVal with MakesAnyKeyStepper[K] {
+  @inline def keyStepper: AnyStepper[K] with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, HE](underlying)
+    new StepsAnyHashTableKey(tbl, 0, tbl.length)
+  }
+}
+
+final class RichHashTableDoubleKeyCanStep[HE >: Null <: collection.mutable.HashEntry[Double, HE]](private val underlying: collection.mutable.HashTable[Double, HE])
+extends AnyVal with MakesDoubleKeyStepper {
+  @inline def keyStepper: DoubleStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[Double, HE](underlying)
+    new StepsDoubleHashTableKey(tbl, 0, tbl.length)
+  }
+}
+
+final class RichHashTableIntKeyCanStep[HE >: Null <: collection.mutable.HashEntry[Int, HE]](private val underlying: collection.mutable.HashTable[Int, HE])
+extends AnyVal with MakesIntKeyStepper {
+  @inline def keyStepper: IntStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[Int, HE](underlying)
+    new StepsIntHashTableKey(tbl, 0, tbl.length)
+  }
+}
+
+final class RichHashTableLongKeyCanStep[HE >: Null <: collection.mutable.HashEntry[Long, HE]](private val underlying: collection.mutable.HashTable[Long, HE])
+extends AnyVal with MakesLongKeyStepper {
+  @inline def keyStepper: LongStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[Long, HE](underlying)
+    new StepsLongHashTableKey(tbl, 0, tbl.length)
+  }
+}
+
+// Steppers for entries stored in DefaultEntry HashEntry
+// (both for key-value pair and for values alone)
+
+final class RichDefaultHashTableCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, V]])
+extends AnyVal with MakesAnyStepper[(K, V)] {
+  @inline def stepper: AnyStepper[(K,V)] with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, V]](underlying)
+    new StepsAnyDefaultHashTable(tbl, 0, tbl.length)
+  }
+}
+
+final class RichDefaultHashTableValueCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, V]])
+extends AnyVal with MakesAnyValueStepper[V] {
+  @inline def valueStepper: AnyStepper[V] with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, V]](underlying)
+    new StepsAnyDefaultHashTableValue(tbl, 0, tbl.length)
+  }
+}
+
+final class RichDefaultHashTableDoubleValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, Double]])
+extends AnyVal with MakesDoubleValueStepper {
+  @inline def valueStepper: DoubleStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, Double]](underlying)
+    new StepsDoubleDefaultHashTableValue(tbl, 0, tbl.length)
+  }
+}
+
+final class RichDefaultHashTableIntValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, Int]])
+extends AnyVal with MakesIntValueStepper {
+  @inline def valueStepper: IntStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, Int]](underlying)
+    new StepsIntDefaultHashTableValue(tbl, 0, tbl.length)
+  }
+}
+
+final class RichDefaultHashTableLongValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.DefaultEntry[K, Long]])
+extends AnyVal with MakesLongValueStepper {
+  @inline def valueStepper: LongStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.DefaultEntry[K, Long]](underlying)
+    new StepsLongDefaultHashTableValue(tbl, 0, tbl.length)
+  }
+}
+  
+// Steppers for entries stored in LinkedEntry HashEntry
+// (both for key-value pair and for values alone)
+
+final class RichLinkedHashTableCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, V]])
+extends AnyVal with MakesAnyStepper[(K,V)] {
+  @inline def stepper: AnyStepper[(K,V)] with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, V]](underlying)
+    new StepsAnyLinkedHashTable(tbl, 0, tbl.length)
+  }
+}
+
+final class RichLinkedHashTableValueCanStep[K, V](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, V]])
+extends AnyVal with MakesAnyValueStepper[V] {
+  @inline def valueStepper: AnyStepper[V] with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, V]](underlying)
+    new StepsAnyLinkedHashTableValue(tbl, 0, tbl.length)
+  }
+}
+
+final class RichLinkedHashTableDoubleValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, Double]])
+extends AnyVal with MakesDoubleValueStepper {
+  @inline def valueStepper: DoubleStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, Double]](underlying)
+    new StepsDoubleLinkedHashTableValue(tbl, 0, tbl.length)
+  }
+}
+
+final class RichLinkedHashTableIntValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, Int]])
+extends AnyVal with MakesIntValueStepper {
+  @inline def valueStepper: IntStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, Int]](underlying)
+    new StepsIntLinkedHashTableValue(tbl, 0, tbl.length)
+  }
+}
+
+final class RichLinkedHashTableLongValueCanStep[K](private val underlying: collection.mutable.HashTable[K, collection.mutable.LinkedEntry[K, Long]])
+extends AnyVal with MakesLongValueStepper {
+  @inline def valueStepper: LongStepper with EfficientSubstep = {
+    val tbl = CollectionInternals.getTable[K, collection.mutable.LinkedEntry[K, Long]](underlying)
+    new StepsLongLinkedHashTableValue(tbl, 0, tbl.length)
+  }
+}
