+++ src/main/scala/scala/compat/java8/collectionImpl/StepsIterator.scala
@@ -7,6 +7,10 @@ import scala.compat.java8.runtime._
 
 import Stepper._
 
+/////////////////////////////
+// Stepper implementations //
+/////////////////////////////
+
 private[java8] class StepsAnyIterator[A](_underlying: Iterator[A])
 extends StepsLikeIterator[A, StepsAnyIterator[A]](_underlying) {
   def semiclone() = new StepsAnyIterator(null)
@@ -30,3 +34,24 @@ extends StepsLongLikeIterator[StepsLongIterator](_underlying) {
   def semiclone() = new StepsLongIterator(null)
   def nextLong() = if (proxied ne null) proxied.nextStep else underlying.next
 }
+
+//////////////////////////
+// Value class adapters //
+//////////////////////////
+
+final class RichIteratorCanStep[A](private val underlying: Iterator[A]) extends AnyVal with MakesAnySeqStepper[A] {
+  @inline def stepper: AnyStepper[A] = new StepsAnyIterator[A](underlying)
+}
+
+final class RichDoubleIteratorCanStep(private val underlying: Iterator[Double]) extends AnyVal with MakesDoubleSeqStepper {
+  @inline def stepper: DoubleStepper = new StepsDoubleIterator(underlying)
+}
+
+final class RichIntIteratorCanStep(private val underlying: Iterator[Int]) extends AnyVal with MakesIntSeqStepper {
+  @inline def stepper: IntStepper = new StepsIntIterator(underlying)
+}
+
+final class RichLongIteratorCanStep(private val underlying: Iterator[Long]) extends AnyVal with MakesLongSeqStepper {
+  @inline def stepper: LongStepper = new StepsLongIterator(underlying)
+}
+
